Updated README and concatenated functions to reduces number of files.
