+++ todolist
@@ -1,0 +1,5 @@
+1) README - how to compile, how to run, how to use
+2) Organize Files
+3) Check to see if NULL is working
+4) Betty
+5) Buffer
